requirements.txt
# Core crawling & extraction
scrapy>=2.11
trafilatura>=1.9.0
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0
tldextract>=5.1.0
lxml>=5.3.0
requests>=2.32.0
PySocks>=1.7.1
warcio>=1.7.4

# Text/ML
sentence-transformers>=3.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
scipy>=1.11.0
hdbscan>=0.8.36
umap-learn>=0.5.6
networkx>=3.2.0
datasketch>=1.6.5
# Choose one vector store for MVP; start with FAISS CPU
faiss-cpu>=1.8.0

# Utilities
pydantic>=2.7.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
qdrant-client>=1.15.0
google-genai>=0.6.0
stem>=1.8.2
